requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
Pillow==9.0.0
